+++ example/subscriber.dart
@@ -12,7 +12,7 @@ void main() async {
   });
 
   await stan.connect(host: 'localhost', clusterID: 'default');
-  final subscription = await stan.subscribe(subject: 'test', queueGroup: 'subscriber', durableName: 'fred');
+  final subscription = await stan.subscribe(subject: 'test', queueGroup: 'queue-name', durableName: 'durable-name');
   await for (final dataMessage in subscription!.stream) {
     print(dataMessage.asString());
     stan.acknowledge(subscription, dataMessage);